+++ packages/components/input/src/Input.tsx
@@ -17,8 +17,19 @@ export interface InputProps extends InputPrimitiveProps {
   onValueChange?: (value: string) => void
 }
 
-export const Input = forwardRef<HTMLInputElement, InputProps>(
-  ({ className, asChild, onValueChange, onChange, onKeyDown, ...others }, ref) => {
+const Root = forwardRef<HTMLInputElement, InputProps>(
+  (
+    {
+      className,
+      asChild = false,
+      onValueChange,
+      onChange,
+      onKeyDown,
+      disabled: disabledProp,
+      ...others
+    },
+    ref
+  ) => {
     const field = useFormFieldControl()
     const group = useInputGroup()
 
@@ -33,6 +44,7 @@ export const Input = forwardRef<HTMLInputElement, InputProps>(
     } = group
     const Component = asChild ? Slot : 'input'
     const state = field.state ?? group.state
+    const disabled = group.disabled ?? disabledProp
 
     const handleChange: ChangeEventHandler<HTMLInputElement> = event => {
       if (onChange) {
@@ -60,6 +72,7 @@ export const Input = forwardRef<HTMLInputElement, InputProps>(
         id={id}
         name={name}
         className={inputStyles({
+          asChild,
           className,
           intent: state,
           hasLeadingAddon: !!hasLeadingAddon,
@@ -68,7 +81,7 @@ export const Input = forwardRef<HTMLInputElement, InputProps>(
           hasTrailingIcon: !!hasTrailingIcon,
           hasClearButton: !!hasClearButton,
         })}
-        disabled={group.disabled}
+        disabled={disabled}
         required={isRequired}
         aria-describedby={description}
         aria-invalid={isInvalid}
@@ -79,5 +92,9 @@ export const Input = forwardRef<HTMLInputElement, InputProps>(
     )
   }
 )
+
+export const Input = Object.assign(Root, {
+  id: 'Input',
+})
 
 Input.displayName = 'Input'