+++ packages/components/textarea/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.5.1](https://github.com/adevinta/spark/compare/@spark-ui/textarea@0.5.0...@spark-ui/textarea@0.5.1) (2023-07-26)
+
+**Note:** Version bump only for package @spark-ui/textarea
+
+# [0.5.0](https://github.com/adevinta/spark/compare/@spark-ui/textarea@0.4.2...@spark-ui/textarea@0.5.0) (2023-07-26)
+
+### Features
+
+- **textarea:** improve textarea props and doc ([af8cce2](https://github.com/adevinta/spark/commit/af8cce23ce98240549cb71e4ce4ec4e8ebfe7d99))
+- **textarea:** update textarea ([efab349](https://github.com/adevinta/spark/commit/efab3499ecb7c46a24e856a0bb5f211f85df6eda))
+- **textarea:** update using id ([dfc2d29](https://github.com/adevinta/spark/commit/dfc2d293153a20282a3f8f90208c4c9d21e794e9))
+
+## [0.4.2](https://github.com/adevinta/spark/compare/@spark-ui/textarea@0.4.1...@spark-ui/textarea@0.4.2) (2023-07-26)
+
+**Note:** Version bump only for package @spark-ui/textarea
+
 ## [0.4.1](https://github.com/adevinta/spark/compare/@spark-ui/textarea@0.4.0...@spark-ui/textarea@0.4.1) (2023-07-24)
 
 **Note:** Version bump only for package @spark-ui/textarea